requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-auth==2.17.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
idna==3.4
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
